Use actor id when querying from revision_userindex
The rev_user field of the revision_userindex table is no longer indexed,
see [0] [1].  Instead the rev_actor field, a reference to the actor
table, should be used.  To speed up the queries, this patch replaces the
user_id with the actor_id and uses rev_actor instead of rev_user when
querying the revision_userindex table.

[0] https://phabricator.wikimedia.org/T223406
[1] https://phabricator.wikimedia.org/T224663